Moved ImageCommentLoaderMainQueueDispatchDecorator to it's own file
